+++ check_manifest.py
@@ -37,7 +37,7 @@ except ImportError:
     import configparser as ConfigParser
 
 
-__version__ = '0.12.dev0'
+__version__ = '0.12'
 __author__ = 'Marius Gedminas <marius@gedmin.as>'
 __licence__ = 'GPL v2 or later' # or ask me for MIT
 __url__ = 'https://github.com/mgedmin/check-manifest'